Move job_control to file under tui module, fix results
