+++ codex-rs/tui/src/tui.rs
@@ -1,8 +1,8 @@
 use std::fmt;
 use std::io::IsTerminal;
+use std::io::Result;
 use std::io::Stdout;
 use std::io::stdout;
-use std::io::{self};
 use std::panic;
 use std::pin::Pin;
 use std::sync::Arc;
@@ -42,10 +42,13 @@ use crate::tui::job_control::SUSPEND_KEY;
 #[cfg(unix)]
 use crate::tui::job_control::SuspendContext;
 
+#[cfg(unix)]
+mod job_control;
+
 /// A type alias for the terminal type used in this application
 pub type Terminal = CustomTerminal<CrosstermBackend<Stdout>>;
 
-pub fn set_modes() -> io::Result<()> {
+pub fn set_modes() -> Result<()> {
     execute!(stdout(), EnableBracketedPaste)?;
 
     enable_raw_mode()?;
@@ -77,8 +80,8 @@ impl Command for EnableAlternateScroll {
     }
 
     #[cfg(windows)]
-    fn execute_winapi(&self) -> io::Result<()> {
-        Err(io::Error::other(
+    fn execute_winapi(&self) -> Result<()> {
+        Err(std::io::Error::other(
             "tried to execute EnableAlternateScroll using WinAPI; use ANSI instead",
         ))
     }
@@ -98,8 +101,8 @@ impl Command for DisableAlternateScroll {
     }
 
     #[cfg(windows)]
-    fn execute_winapi(&self) -> io::Result<()> {
-        Err(io::Error::other(
+    fn execute_winapi(&self) -> Result<()> {
+        Err(std::io::Error::other(
             "tried to execute DisableAlternateScroll using WinAPI; use ANSI instead",
         ))
     }
@@ -112,7 +115,7 @@ impl Command for DisableAlternateScroll {
 
 /// Restore the terminal to its original state.
 /// Inverse of `set_modes`.
-pub fn restore() -> io::Result<()> {
+pub fn restore() -> Result<()> {
     // Pop may fail on platforms that didn't support the push; ignore errors.
     let _ = execute!(stdout(), PopKeyboardEnhancementFlags);
     execute!(stdout(), DisableBracketedPaste)?;
@@ -123,9 +126,9 @@ pub fn restore() -> io::Result<()> {
 }
 
 /// Initialize the terminal (inline viewport; history stays in normal scrollback)
-pub fn init() -> io::Result<Terminal> {
+pub fn init() -> Result<Terminal> {
     if !stdout().is_terminal() {
-        return Err(io::Error::other("stdout is not a terminal"));
+        return Err(std::io::Error::other("stdout is not a terminal"));
     }
     set_modes()?;
 
@@ -306,7 +309,7 @@ impl Tui {
 
     /// Enter alternate screen and expand the viewport to full terminal size, saving the current
     /// inline viewport for restoration when leaving.
-    pub fn enter_alt_screen(&mut self) -> io::Result<()> {
+    pub fn enter_alt_screen(&mut self) -> Result<()> {
         let _ = execute!(self.terminal.backend_mut(), EnterAlternateScreen);
         // Enable "alternate scroll" so terminals may translate wheel to arrows
         let _ = execute!(self.terminal.backend_mut(), EnableAlternateScroll);
@@ -325,7 +328,7 @@ impl Tui {
     }
 
     /// Leave alternate screen and restore the previously saved inline viewport, if any.
-    pub fn leave_alt_screen(&mut self) -> io::Result<()> {
+    pub fn leave_alt_screen(&mut self) -> Result<()> {
         // Disable alternate scroll when leaving alt-screen
         let _ = execute!(self.terminal.backend_mut(), DisableAlternateScroll);
         let _ = execute!(self.terminal.backend_mut(), LeaveAlternateScreen);
@@ -345,7 +348,7 @@ impl Tui {
         &mut self,
         height: u16,
         draw_fn: impl FnOnce(&mut custom_terminal::Frame),
-    ) -> io::Result<()> {
+    ) -> Result<()> {
         // If we are resuming from ^Z, we need to prepare the resume action now so we can apply it
         // in the synchronized update.
         #[cfg(unix)]
@@ -490,8 +493,8 @@ impl Command for PostNotification {
     }
 
     #[cfg(windows)]
-    fn execute_winapi(&self) -> io::Result<()> {
-        Err(io::Error::other(
+    fn execute_winapi(&self) -> Result<()> {
+        Err(std::io::Error::other(
             "tried to execute PostNotification using WinAPI; use ANSI instead",
         ))
     }
@@ -499,191 +502,5 @@ impl Command for PostNotification {
     #[cfg(windows)]
     fn is_ansi_code_supported(&self) -> bool {
         true
-    }
-}
-
-#[cfg(unix)]
-mod job_control {
-    use std::io;
-    use std::io::stdout;
-    use std::sync::Arc;
-    use std::sync::Mutex;
-    use std::sync::PoisonError;
-    use std::sync::atomic::AtomicBool;
-    use std::sync::atomic::AtomicU16;
-    use std::sync::atomic::Ordering;
-
-    use crossterm::cursor::MoveTo;
-    use crossterm::cursor::Show;
-    use crossterm::event::KeyCode;
-    use crossterm::terminal::EnterAlternateScreen;
-    use crossterm::terminal::LeaveAlternateScreen;
-    use ratatui::crossterm::execute;
-    use ratatui::layout::Position;
-    use ratatui::layout::Rect;
-
-    use crate::key_hint;
-
-    use super::DisableAlternateScroll;
-    use super::EnableAlternateScroll;
-    use super::Terminal;
-
-    pub(crate) const SUSPEND_KEY: key_hint::KeyBinding = key_hint::ctrl(KeyCode::Char('z'));
-
-    /// Coordinates suspend/resume handling so the TUI can restore terminal context after SIGTSTP.
-    ///
-    /// On suspend, it records which resume path to take (realign inline viewport vs. restore alt
-    /// screen) and caches the inline cursor row so the cursor can be placed meaningfully before
-    /// yielding.
-    ///
-    /// After resume, `prepare_resume_action` consumes the pending intent and returns a
-    /// `PreparedResumeAction` describing any viewport adjustments to apply inside the synchronized
-    /// draw.
-    ///
-    /// Callers keep `suspend_cursor_y` up to date during normal drawing so the suspend step always
-    /// has the latest cursor position.
-    ///
-    /// The type is `Clone`, using Arc/atomic internals so bookkeeping can be shared across tasks
-    /// and moved into the boxed `'static` event stream without borrowing `self`.
-    #[derive(Clone)]
-    pub(crate) struct SuspendContext {
-        /// Resume intent captured at suspend time; cleared once applied after resume.
-        resume_pending: Arc<Mutex<Option<ResumeAction>>>,
-        /// Inline viewport cursor row used to place the cursor before yielding during suspend.
-        suspend_cursor_y: Arc<AtomicU16>,
-    }
-
-    impl SuspendContext {
-        pub(crate) fn new() -> Self {
-            Self {
-                resume_pending: Arc::new(Mutex::new(None)),
-                suspend_cursor_y: Arc::new(AtomicU16::new(0)),
-            }
-        }
-
-        /// Capture how to resume, stash cursor position, and temporarily yield during SIGTSTP.
-        ///
-        /// - If the alt screen is active, exit alt-scroll/alt-screen and record `RestoreAlt`;
-        ///   otherwise record `RealignInline`.
-        /// - Update the cached inline cursor row so suspend can place the cursor meaningfully.
-        /// - Trigger SIGTSTP so the process can be resumed and continue drawing with the saved state.
-        pub(crate) fn suspend(&self, alt_screen_active: &Arc<AtomicBool>) -> io::Result<()> {
-            if alt_screen_active.load(Ordering::Relaxed) {
-                // Leave alt-screen so the terminal returns to the normal buffer while suspended; also turn off alt-scroll.
-                let _ = execute!(stdout(), DisableAlternateScroll);
-                let _ = execute!(stdout(), LeaveAlternateScreen);
-                self.set_resume_action(ResumeAction::RestoreAlt);
-            } else {
-                self.set_resume_action(ResumeAction::RealignInline);
-            }
-            let y = self.suspend_cursor_y.load(Ordering::Relaxed);
-            let _ = execute!(stdout(), MoveTo(0, y), Show);
-            suspend_process()
-        }
-
-        /// Consume the pending resume intent and precompute any viewport changes needed post-resume.
-        ///
-        /// Returns a `PreparedResumeAction` describing how to realign the viewport once drawing
-        /// resumes; returns `None` when there was no pending suspend intent.
-        pub(crate) fn prepare_resume_action(
-            &self,
-            terminal: &mut Terminal,
-            alt_saved_viewport: &mut Option<Rect>,
-        ) -> Option<PreparedResumeAction> {
-            let action = self.take_resume_action()?;
-            match action {
-                ResumeAction::RealignInline => {
-                    let cursor_pos = terminal
-                        .get_cursor_position()
-                        .unwrap_or(terminal.last_known_cursor_pos);
-                    let viewport = Rect::new(0, cursor_pos.y, 0, 0);
-                    Some(PreparedResumeAction::RealignViewport(viewport))
-                }
-                ResumeAction::RestoreAlt => {
-                    if let Ok(Position { y, .. }) = terminal.get_cursor_position()
-                        && let Some(saved) = alt_saved_viewport.as_mut()
-                    {
-                        saved.y = y;
-                    }
-                    Some(PreparedResumeAction::RestoreAltScreen)
-                }
-            }
-        }
-
-        /// Set the cached inline cursor row so suspend can place the cursor meaningfully.
-        ///
-        /// Call during normal drawing when the inline viewport moves so suspend has a fresh cursor
-        /// position to restore before yielding.
-        pub(crate) fn set_cursor_y(&self, value: u16) {
-            self.suspend_cursor_y.store(value, Ordering::Relaxed);
-        }
-
-        /// Record a pending resume action to apply after SIGTSTP returns control.
-        fn set_resume_action(&self, value: ResumeAction) {
-            *self
-                .resume_pending
-                .lock()
-                .unwrap_or_else(PoisonError::into_inner) = Some(value);
-        }
-
-        /// Take and clear any pending resume action captured at suspend time.
-        fn take_resume_action(&self) -> Option<ResumeAction> {
-            self.resume_pending
-                .lock()
-                .unwrap_or_else(PoisonError::into_inner)
-                .take()
-        }
-    }
-
-    /// Captures what should happen when returning from suspend.
-    ///
-    /// Either realign the inline viewport to keep the cursor position, or re-enter the alt screen
-    /// to restore the overlay UI.
-    #[derive(Copy, Clone, Debug, Eq, PartialEq)]
-    pub(crate) enum ResumeAction {
-        /// Shift the inline viewport to keep the cursor anchored after resume.
-        RealignInline,
-        /// Re-enter the alt screen and restore the overlay UI.
-        RestoreAlt,
-    }
-
-    /// Describes the viewport change to apply when resuming from suspend during the synchronized draw.
-    ///
-    /// Either restore the alt screen (with viewport reset) or realign the inline viewport.
-    #[derive(Clone, Debug)]
-    pub(crate) enum PreparedResumeAction {
-        /// Re-enter the alt screen and reset the viewport to the terminal dimensions.
-        RestoreAltScreen,
-        /// Apply a viewport shift to keep the inline cursor position stable.
-        RealignViewport(Rect),
-    }
-
-    impl PreparedResumeAction {
-        pub(crate) fn apply(self, terminal: &mut Terminal) -> io::Result<()> {
-            match self {
-                PreparedResumeAction::RealignViewport(area) => {
-                    terminal.set_viewport_area(area);
-                }
-                PreparedResumeAction::RestoreAltScreen => {
-                    execute!(terminal.backend_mut(), EnterAlternateScreen)?;
-                    // Enable "alternate scroll" so terminals may translate wheel to arrows
-                    execute!(terminal.backend_mut(), EnableAlternateScroll)?;
-                    if let Ok(size) = terminal.size() {
-                        terminal.set_viewport_area(Rect::new(0, 0, size.width, size.height));
-                        terminal.clear()?;
-                    }
-                }
-            }
-            Ok(())
-        }
-    }
-
-    /// Deliver SIGTSTP after restoring terminal state, then re-applies terminal modes once resumed.
-    fn suspend_process() -> io::Result<()> {
-        super::restore()?;
-        unsafe { libc::kill(0, libc::SIGTSTP) };
-        // After the process resumes, reapply terminal modes so drawing can continue.
-        super::set_modes()?;
-        Ok(())
     }
 }
